Move back tensorflow-related functions
