+++ src/common/model_utils/tests/test_preprocessing.py
@@ -6,39 +6,8 @@ import tensorflow as tf
 
 sys.path.append(str(Path(__file__).parents[1]))
 
-from config import DATA_AUGMENTATION_SAME_PER_CHANNEL, DATA_AUGMENTATION_DIFFERENT_EACH_CHANNEL, DATA_AUGMENTATION_NO  # noqa: E402
 from preprocessing import augmentation, tf_augment_sample, sample_systematic_from_artifacts, sample_windows_from_artifacts, REGEX_PICKLE  # noqa: E402
 
-
-def test_tf_augment_sample():
-    # Test with dataset (not eager)
-    sample = tf.random.uniform((240, 180, 5))
-    target = tf.constant([92.3])
-    dataset = tf.data.Dataset.from_tensors((sample, target))
-    _ = dataset.map(tf_augment_sample, tf.data.experimental.AUTOTUNE)
-
-    # Test eagerly
-    tf_augment_sample(sample, target)
-
-
-def test_imgaug_on_multichannel_same():
-    sample = np.ones((240, 180, 5)) * 0.5
-    result = augmentation(sample, mode=DATA_AUGMENTATION_SAME_PER_CHANNEL)
-    # assert np.all(result[0] == result[1])  # cannot be ensured currently
-    assert result.shape == (240, 180, 5)
-
-
-def test_imgaug_on_multichannel_different():
-    sample = np.ones((240, 180, 5)) * 0.5
-    result = augmentation(sample, mode=DATA_AUGMENTATION_DIFFERENT_EACH_CHANNEL)
-    assert not np.all(result[0] == result[1])
-    assert result.shape == (240, 180, 5)
-
-
-def test_imgaug_on_multichannel_no():
-    sample = np.random.rand(240, 180, 5)
-    result = augmentation(sample, mode=DATA_AUGMENTATION_NO)
-    assert result.shape == (240, 180, 5)
 
 
 def test_sample_windows_from_artifacts_multiple_results():
